Fix IPFS swarm addresses for Bacalhau-run networks.
We've been using the wrong peering addresses for several months!! In
our production environment, we specify the IPFS peering addresses as:
https://github.com/bacalhau-project/bacalhau/blob/main/pkg/config/configenv/production.go#L54-L58
These are correct IP addresses and ports, but… the P2P keys are not
correct. This is actually the P2P key of the Bacalhau node and not
the IPFS node!

A correct ID should look like:

/ip4/35.245.115.191/tcp/4001/p2p/12D3KooWE4wfAknWtY9mQ4eAA8zrFGeZa7X2Kh4nBP2tZgDSt7Rh

So when our clients try and connect to the Bacalhau IPFS nodes, they
all fail to connect because they are expecting to see a different P2P
key.

And due to an unfortunate logging bug, all of the error messages that
get printed are not visible even when using LOG_LEVEL=debug so noone
has noticed all of these peer connection errors.

What about the times where we've successfully downloaded? Well, that's
because we've still been connecting to the public bootstrap IPFS
nodes. Job results eventually make it onto the public nodes. This is
why you can run a job and not be able to get the results, but
magically the results appear some time later.So the solution for the
--private-internal-ipfs issue is actually to fix our config. When we
have the correct peer addresses, everything works fine!

These were fetched using a command like:

    gcloud compute instances list \
        --project=bacalhau-staging \
        --filter='name~bacalhau-vm' \
        --format='value(name)' |
    xargs -n1 gcloud compute ssh \
        --project=bacalhau-staging \
        --command='bash -c "sudo env IPFS_PATH=/data/ipfs ipfs id"' \
        --zone=us-east4-c |
    jq -r '.Addresses|.[]' | tee staging.txt