Check for running containing on docker start
When using the docker executor, the start function checks its map to see
if it has any containers currently running for the provided execution.
This will avoid dupes.  During a restart however, that map is empty and
so the only way to check if there is a running docker container, is to
look for it, which this PR attempts to do.