Fix race condition in noop executor
Previously the code was closing the done channel, and then setting the
result. Unfortunately, there is a goroutine waiting for the done
channel to close to then process the result. Occassionally this fails in
tests with a nil pointer error similar as it was doing this...

```go
// in handler
close(e.done)
e.result = {...}

// in executor
case <-handler.done:
     ...
     if handler.result.err != nil
```