Provide option to turn on HTTPS for SDK
Currently the use of http is hard-coded.  This minimal change allows the
user to specify the use of HTTPS by setting the `BACALHAU_HTTPS`
environment variable - leaving the default as http to avoid breakage.

When breaking changes make sense, we should move to specifying the
scheme as part of the API_HOST and default to https, but for now this
ensures users can continue to use the SDK without immediate breakage.