Allow different state counts in flaky test
The `publish-partial-failure` test, expects the end result of the test
to be Fail, Fail, Cancelled. However, sometimes, the test expecting to
be cancelled actually completes in time, leaving us with Fail, Fail,
Complete. Any attempt to Cancel the completed task will fail as it is a
bad state transition, and the object revision will have changed leaving
the call to Cancel to fail with a revision error from the job store.

To "resolve" this, we instead allow the result for this to to be one of:

* Fail, Fail, Complete
* Fail, Fail, Cancel

In the absence of a usable Range type, the test introduces an IntMatch
type which contains a simple list of numbers allowing us to later
provide a single int in the Match method to see if it exists in the list
of integers.  We _could_ have achieved the same with []int{} throughout
the test cases, but this felt a little cleaner (and more testable).