Use non-p2p multiaddress for peer
Currently we need to use a P2P multiaddress at compute node startup,
which requires that is know the ip address, the ip port and the peer id
of the requester node.

This PR allows the user to specify a HTTP or HTTPS multiaddress (e.g.
/tcp/127.0.0.1/ip/1234/http) that points to the requester node. This
address is then used to call the node info endpoint to retrieve the
peerid, before using it to construct a p2p multiaddress.