Limit update checks to once per day
In #2893 we made more commands check for updates. Strictly not all of
these checks are necessary as the code is unlikely to have changed
regularly. Instead we should limit the amount we check for updates to
(perhaps) once per day.

Every time we perform an update check, we write the current time into
update.json. Every time we run a command. if the current time is less
than 24h after this value, don't do any checking. If the current time
is 24h ahead of this value, do an update check.