disable execution timeout for long running jobs (#2946)
This PR fixes an issue where default execution timeout is still applied
to long running jobs in the compute side, causing them to fail. While no
timeout is applied and no result is expected, the existing go routine
that is waiting is still there so that it can release the allocated
resources in case of failure or cancellation.