Adds SourcePath in conversion to legacy spec
Currently if you run a job with

```
bacalhau docker run -i "src=file://path,dst=/inputs" ...
```

when you call describe on that job, the Inputs contains an entry where
StorageSource = localDirectory, but copying that spec into a yaml file
and trying to pipe it through `bacalhau create` fails because of a
missing SourcePath attribute in the StorageSource.

This PR ensures when converting from a new format spec to a legacy
spec, we include the SourcePath as `SourcePath` and not `Path` as
previously (which was still being overwritten with the dst value).