fix only one job is restarted during startup (#2979)
What we observe today is during node startup, only a single long running
job is restarted. When you try to stop/cancel that job, then the next
one restarts, and so on. The problem is with our execution deque backoff
logic.

Every time an execution is submitted or finishes, we call `deque` which
looks for the oldest execution that is eligible to run based on
available and requested capacity. We had a backoff logic configured to
50ms to minimize the number of queue iterations during high throughput.
However, this was a premature optimization and can cause starvation with
long running jobs as they don't finish!

The issue during startup:
1. `Run` Execution1 which enqueue the execution
2. `deque` polls Execution1 and actually starts it
3. `deque` updates backoff to 50ms in the future
4. `Run` Execution2 which enqueue the execution
5. `deque` returns because of backoff logic

The fix is to remove backoff logic. In the future, we can look into a
better implementation for dequeuing executions or switching to a
periodic background task instead of each time a job is submitted, when
this becomes a problem