Changes how we choose a log stream address
Currently when the requester node wants to read the logs from a compute
node, it asks the compute node which responds with the first TCP address
it can find (this may be ip4 or ip6). If it fails to do this, it will
instead just use the first address in the list of addresses for the
host.

We are encountering situations where a loopback address is being chosen
in preference to a private network address which is preferable in nearly
all cases (except devstack).

This PR attempts to order the available addresses into a more effective
ordering. The addresses that are chosen from are ordered by:

* Private addresses (from RFC1918)
* Public addresses
* Loopback addresses (e.g. 127.0.0.1)
* Unspecified addresses (such as ip6 ::)
* Link local addresses (from RFC3927)