Show node ranking failures in the CLI (again)
We used to show a list of nodes that had failed ranking when not
enough nodes were found to run the job. At some point we regressed
and no longer showed those errors, leading to lots of confusion
around why jobs aren't running. This commit restores the view of node
ranking failures being passed to the CLI.