Allow ipfs to run in devstack when ./plugins exists
Currently at startup, if kubo/ipfs finds a plugins folder inside the
repo, it will crash if any binary found in that directory (or children)
is not a kubo plugin.

When we start ipfs as part of devstack, we do not specify a repo and so
it means we are unable to have a ./plugins folder in the directory where
we run devstack.

This PR creates a temp directory for devstack's IPFS to pretend is a
repo so that we can put a plugins folder where we like.