Initial layout of plugins directory
Lays out the structure for a plugins directory, including Makefiles
where appropriate which will be able to build the plugins once they are
moved here.  Currently the only functionality implemented is the new
home for the protobuf, and a makefile to build it.  This task will need
to move into the executors own directories once their own build infra is
in place.

This PR is blocked by
https://github.com/bacalhau-project/bacalhau/pull/3024 which is blocked
by CI configuration.