Temporarily disable uploading of build metadata
The build previously wrote a JSON dict containing
`x-goog-meta-last-updated` to a file in a GS bucket.  This lived in the
CI/CD account which is now deprecated.

This PR disables the uploading of this metadata pending clarification of
its use, and whether it is necessary.  If so it can be re-enabled once
the appropriate storage and creds have been obtained.