Layout of plugins directory (#3026)
Lays out the structure for a plugins directory, including Makefiles
where appropriate which will be able to build the plugins once they are
moved here. Currently the only functionality implemented is the new home
for the protobuf, and a makefile to build it. This task will need to
move into the executors own directories once their own build infra is in
place.

Co-authored-by: Forrest <6546409+frrist@users.noreply.github.com>