Extract wasm executor into standalone location
Moves the key parts of the wasm executor to the plugins folder, and
removes some key references back to the main bacalhau packages.

Adds more useful types to the protobufs defining the interface for
remote executors.

Unfortunately still has deep dependency on the core bacalhau packages
that are hard to remove.