support GCP buckets (#3040)
This PR adds support for S3 compatible buckets, such as Google Buckets. 

The problem is with `Accept-Encoding` not supported by GCP, but AWS SDK
v2 is including that header in the request signature. The solution is to
add a middleware that removes the header before the request is signed,
and then put it back in.

The second problem is with the `publisher` where it passes
`ChecksumAlgorithmSha256` header to create a checksum for the object,
but this is not supported by GCP. The solution now is to only pass this
flag if we are publishing to an AWS S3 bucket.

### Testing
```YAML
Type: batch
Count: 1
Tasks:
  - Name: main
    Engine:
      Type: docker
      Params:
        Image: ubuntu:latest
        Entrypoint:
          - /bin/bash
        Parameters:
          - -c
          - ls /input_custom
    Publisher:
      Type: s3
      Params:
        Bucket: walid-test-devboxes
        Key: publisher/walid-manual-test-{executionID}
        Endpoint: https://storage.googleapis.com
        Compress: true
    InputSources:
      - Alias: input_custom
        Target: /input_custom
        Source:
          Type: s3
          Params:
            Bucket: walid-test-devboxes
            Key: set1/
            Endpoint: https://storage.googleapis.com
```

This job is running as expected where the bucket is a GCP bucket. Also
verified that `bacalhau get <job_id>` is working and a pre-signed url is
generated correctly.

### Requirements
You still need to have validate [GCP HMAC
credentials](https://cloud.google.com/storage/docs/authentication/managing-hmackeys)
in your compute nodes. Methods documented here still work, expect for
IAM roles
https://docs.bacalhau.org/references/other-specifications/sources/s3#credential-requirements