Serve the Web UI with correct host/port from Bacalhau
This commit upgrades the `--web-ui` flag to serve the static React app
from an embedded filesystem inside Bacalhau. This means that servers
running nodes do not need any more components available to also
expose the dashboard.

We do this by turning the index.html into a Go template, and then
subbing in the correct host and port at boot of the binary. By default
we serve on port 80 and this isn't configurable yet.

Assuming the Node.js infrastructure is installed, do a first run of
`make init` and then simply run `make build` as usual to build the web
app along with the binary if necessary.