Fix StorageProviders to store where they are told
Currently storage providers create a temporary folder underneath the
system configured storage path, where the folder include the provider's
name.  This means all of the data is indexed on the provider, rather
than the execution.

This PR switches to telling the provider where to store the data, so
that it is keyed by job/execution instead (also within the configured
data folder).