Remove legacy resource usage config parsing
We have two structures for handling resource configuration which are
completely equivalent. We definitely don't need both of them.

This commit converts the config package to use the new one, allowing
us to remove the parsing for the redundant one. It is still used in
the old job spec and so needs to stick around until that is firmly
deprecated.