Support multiple and also AMD GPUs in Docker jobs
This commit updates our Docker container configuration code to add as
many GPUs to the container as have been allocated to the job.
Previously, only 1 GPU was added regardless of how many were
requested.

It also adds support for AMD GPUs in Docker containers.

Notably, the executor system now allocates the GPUs that will be used
by the container. Because we have to explicitly declare in Docker-land
which GPUs we want to expose, if we didn't do this, we wouldn't be
able to control resource usage of GPUs and any job that requested any
number of GPUs would just be able to access them all.

So it is now the Tracker that allocates any requested but unallocated
resources when it does its tracking. This has the benefit that if/when
we want to support requesting specific GPUs in the job spec, this will
all already work as long as they are populated in the incoming
Resources object.