More gracefully handle bad tool output
It turns out that `rocm-smi` will print non-JSON output but still
return a zero error code in the case of an error – naughty! This has
prompted us to be more uniform with how we handle tool errors – any
error generated by a GPU detector tool will be ignored, rather than
errors that they explicitly flag. This will allow the node to come
online with zero GPUs in this case, and the reason for this will be in
the log.