Tell StorageProviders where to store data (#3053)
Currently storage providers create a temporary folder underneath the
system configured storage path, where the folder include the provider's
name. This means all of the data is indexed on the provider, rather than
the execution. Most of the StorageProvider implementations determine
these paths when they are first created.

This PR switches to telling the provider where to store the data, so
that it is keyed by job/execution instead (also within the configured
data folder). This provides a single location for all inputs affecting
an execution (or all local executions for a Job).