Copy large inline sources to IPFS
Currently an inline attachment sent to the requester endpoint will be
replaced with an IPFS CID if the inline source is above a certain
source.  We want to have the same functionality in the orchestrator
endpoint so that we can submit jobs without concern that we are
inflating the size of the data store unnecessarily.