Adds a very early `bacalhau exec` command
Adds an exec command which can be used for submitting custom job types.
Because these custom job types could be anything, then we have to take
steps to cater for lots of potential edge cases.

This initial version supports duckdb and python, with both templates
currently being embedded in the binary. Ideally we should retrieve these
from somewhere, either on disk or from a requester node (which can serve
the templates based on the translators installed).

The command after exec is used as the Command: value in the Engine
Params.  Remaining parameters that are not bacalhau parameters are
passed as the Engine.Params["Arguments"] string array.

If the user wishes to pass any parameters that clash with bacalhau (say
-i) they must do so after using --

Some commands will need to have code to run.  To support this we have
added --code which allows the user to specify a single file or a folder
of files.  As long as the total size is less than 10Mb the contents will
be added as an inline storage to the job specification to be mounted at
/code with an alias of 'code'.

`bacalhau exec -i src=...,dst=... duckdb "select * from /inputs/data.csv"`

Sends a request where Command: duckdb, Arguments: ["select * from
/inputs/data.csv"].

`bacalhau exec python app.py -x -y --code=.`

Sends a request where Command: python, Arguments: ["app.py", "-x", "-y"]
and where . is added as an inline tgz.