Add support for Intel GPUs
Intel GPUs can be discovered using the `xpu-smi`, which we can now
parse the output of to discover device information.

Unfortunately, this is a two-step process where first we have to look
up the list of GPUs and then we have to query extended information
about them, so this requires some interface changes.

Thankfully, the docker integration is the same as for AMD devices.