Correctly expose only the allocated GPU devices to Docker jobs
Previously we assumed that there was a simple mapping of:

    /dev/dri/cardN -> /dev/dri/renderD$(N + 128)

However by inspection this is not the case. Instead, we have to know
the PCI address of the card and use the symlinks in /dev/dri/by-path
to look up the correct card and render devices.

So we now collect PCI addresses for AMD and Intel cards. We don't need
to do this for Nvidia cards because they are exposed by a different
mechanism.