Remove building Bacalhau as part of Docker image build
This meant that our new Web UI build steps were not being picked up.
This is not really necessary as we can just reuse the output that was
created from our CI build step, rather than the rigmarole of trying to
capture the right files and recreate the build process within the
Dockerfile.

This change also means we can build the docker image on PRs as well,
to check that it works.

There are some other refactors here – we now build and test as
separate jobs so that we don't need to wait for tests to pass before
pushing binaries to a release, and also generate all of our binaries
from a single cross-compiling step (because we don't use CGO and hence
can compile for all platforms from anywhere).