Custom job type docker images
This PR contains images for running the first two custom job types,
python and duckdb.  The images currently build and the python tests work
well (to test the launcher process), but they may require changes in
future in response to the job translation work.

The duckdb image is reasonably straight-forward, it downloads the
appropriate version of duckdb (for the target platform) and makes it
available for use.

The python image is slightly more complex in that it expects code to be
mounted at /code, and a launcher process (launcher.py) is responsible
for moving the core, determining where the requirements are defined, and
then executing the program using the arguments provided by the CLI on
submission.

May require further changes in response to the job translation (and will
require network access for requirements fetching).