Fix the naming convention for jobs and nodes
job Name and ID columns were often showing duplicate
information due to batch jobs not having names.
There is now a simple 'job' column that will show
a shortened ID or Name, depending on whether the job
is of 'batch' type.

For the Nodes page, the name column was unnecessary, so
it has been removed.