feat: further configuration of embedded ipfs daemon (#3096)
### Why
At present, the embedded IPFS node in bacalhau will listen on [random
ports](https://github.com/bacalhau-project/bacalhau/blob/main/pkg/ipfs/node.go#L347).
Users can configure the embedded IPFS node to listen on a "preferred"
swarm address by setting the `BACALHAU_PREFERRED_ADDRESS`, but they are
currently unable to configure which port the node listens on for swarm
connections. **The lack of determinism in setup makes it challenging to
correctly configure firewall rules when deploying bacalhau to
production-like settings, since the port the IPFS node listen on cannot
be known at the time of deployment.**

### What
This PR allows the embedded IPFS nodes gateway, api, and swarm listening
multiaddresses to be configured. Users can configure these values via
the `--ipfs-gateway-listen-addresses`, `--ipfs-api-listen-addresses`,
and `--ipfs-swarm-listen-addresses` flags on the `serve` command
respectively. Alternatively, their correspond environment variables, or
config file value may also be set.
To preserve [backwards
compatibility](https://github.com/bacalhau-project/examples/blob/main/multi-region/tf/node_files/start-bacalhau.sh#L15),
users may continue to use the `BACALHAU_PREFERRED_ADDRESS` environment
variable to configure the swarm address the embedded IPFS listens on,
but should note this variable does not allow the port to be specified,
it will still be randomly assigned.
 
Additionally, this PR makes changes to the behavior of bacalhau and its
embedded IPFS node when the `--ipfs-serve-path` flag is set. The intent
of the flag is to allow users to define a repo location of the embedded
IPFS node to store content and its configuration. The current behavior
deletes this repo when the bacalhau node shuts down. Resulting in any
content stored in the repo being removed, and the identity of the
embedded IPFS node being lost. The new behavior will preserve the
content of the embedded IPFS nodes repo across bacalhau restarts,
maintaining any data the embedded IPFS node stored as well as its
identity. If the `--ipfs-serve-path` flag is not set the behavior
remains unchanged - the repo is considered ephemeral and removed when
bacalhau shuts down.

Lastly, this PR adds an additional flag, `--ipfs-profile`, to configured
the embedded IPFS nodes configuration
[profile](https://github.com/ipfs/kubo/blob/master/docs/config.md#profiles).
The default profiles remain `flatfs` when
`--private-internal-ipfs=false` and `test` when
`--private-internal-ipfs=true`.

---------

Co-authored-by: frrist <forrest@expanso.io>