Show error message when nodes don't support Docker
This commit allows the CLI to show a node ranking error message when
nodes don't have the available engine installed.

We do this by removing the FindNodes method previously used for node
discovery. This used to pre-filter the returned nodes to remove any
that did not support the engine, but this meant that they wouldn't
even appear to the ranking system and wouldn't appear on the CLI
either. This was only here as a performance optimisation, but it's
premature (ranking will still take milliseconds even with a 10,000
node cluster).

Now, we just list all nodes and filter them using the engine ranking
strategy instead. In fact, this strategy has been present for months
and is even enabled, it just never did anything because it was never
passed any nodes to filter out. It works just fine, and provides users
with a helpful error messgae, as below:

    % bacalhau docker run ubuntu date
    Using default tag: latest. Please specify a tag/digest for better reproducibility.
    Job successfully submitted. Job ID: 270ad777-1279-48f1-bc51-c7aa900a90e8
    Checking job status... (Enter Ctrl+C to exit at any time, your job will continue running):

            Communicating with the network  ................  done ✅  0.0s
               Creating job for submission  ................  err  ❌  0.5s

    Error submitting job: not enough nodes to run job. requested: 1, available: 0.
            Node QmT5mqWF: does not support docker, only wasm