Add middleware that runs a Rego policy to determine API access
This commit adds a new `Authorizer` concept that is now used to
control access to APIs. Authorizers have a ShouldAllow method that
receives an inbound HTTP request and says yay or nay to the request
proceeding.

The default implementation of an Authorizer executes policies
implemented in Rego, a policy language. For now, all API servers use
the default `AlwaysAllow` policy which doesn't do any filtering, so
is just a noop.

In a future commit, we will allow the used policies to be customised
through configuration, and implement more authz options through
policies that node operators can choose to install.