Namespace-based policy for anonymous access
This commit adds a policy which aims to match our existing policy as
closely as possible: namely, that write operations on a job namespace
need to be accompanied by an access token, but operations on other
APIs are permitted without any authorization requirement.

We're doing it this way because it's more difficult than it's worth to
match our current scheme exactly, because we don't have access to
primitives to verify a signed request. So instead, we'll need to
implement auth endpoints first before we can activate this.