Enable TLS support for CLI (#2854)
Adds support for making TLS requests to the requester node for both v1
and v2 API clients.

By default, all communication with the requester API will remain HTTP as
switching to HTTPS would be a breaking change. Making TLS calls can be
activated with the command line flag `--tls` or by setting the
environment variable `BACALHAU_API_TLS=1`.

If the server is using a self-signed certificate, it will be necessary
to either accept insecure requests (no certificate verification) or to
provide a CA certificate file.

By specifying --cacert [filename.crt] the client will use a provided CA
certificate to verify the certificate when communicating with the
requester node. This certificate (often called ca.crt) must be provided
by the certificate issuer (which may be the cluster owner in a
self-signed scenario). In addition to the command line flag, it is
possible to set the `BACALHAU_API_CACERT` environment variable to point
to the file.

The least preferable option, but possibly the easiest is to use the
`--insecure` flag to make HTTPS requests without verifying the
certificate. You may also set `BACALHAU_API_INSECURE`.