Store tokens persistently and use them in requests
Tokens are now stored in a simple JSON map in the repo keyed on the
API base URL. This allows the client to communicate with multiple
different requester nodes and keep tokens for each of them.

Tokens are now sent as the Authorization header in API requests, where
they are present. It is not an error for a client to not have a token
because some clusters may continue to accept unauthenticated requests.