Add authentication that supports usernames and passwords
This commit adds `ask` mode authentication that allows node operators
to configure Bacalhau to ask the user for arbitrary information to be
used as a credential.

This method can be used to implement basic usernames and passwords,
shared secrets, security questions and even 2FA.

The associated policy additionally returns a JSON Schema to show what
information is required. The CLI uses the schema to ask the user for
the right information. In the future, the Web UI will do this as well.