Make ExecuteTestCobraCommand safer to use (#3294)
It is currently too easy to forget to add --api-host and --api-port and
then the test may still run, but against the demo network. This will
cause issues later on when the demo network changes.

This PR instead moved the ExecuteTestCobraCommand (and associated)
functions onto the BaseSuite which already specified the host and port.
If the user does not specify api-host or api-port then the configured
values for the suite will be used. This will effectively make it
impossible to run tests against the demo network.

Fixes #3293