Encode useful values in NextToken
NextToken is used to fetch the next page of results, but to do that
safely we need to encode the current sort item, whether it is reversed,
the current limit and offset.  This ensures that we get the expected
results when paging through them.

The CLI will show the next token iff there are more records to fetch.