decouple logsstream from transport layer (#3354)
LogStream is tightly coupled to libp2p to stream logs from the compute
node to the requester, which makes supporting NATS more challenging
https://github.com/bacalhau-project/expanso-planning/issues/413

This PR decouples LogStream and makes fetching logs from the executor
and compute node agnostic to how they are published to the requester.

The way this is achieved by utilizing channels for async vending of
logs. The compute nodes return a channel for
`AsyncResult[models.ExecutionLog]`, which is consumed by the transport
layer (`bprotocol` in current implementation) and sent to the requester.
A similar implementation is also introduced from the client to the
requester node where instead of exposing websockets to the client user,
the new SDK just returns a channel of `AsyncResult[models.ExecutionLog]`

This PR also migrates `logs` endpoint from the old APIs to the new
version, implements logs SDK, and introduces helper methods to expose
and consume additional websockets in the future.