Include supported job type options in help text
Includes the currently supported job types in the help text so that
currently it displays

```
Allows for the execution of a job type with the given code,
without the need to create a container, or webassembly module.
By specifying the code with the '--code' flag you can ship
the code to the cluster for execution, specified by the
remainder of the command line.  See examples below.

 Supported job types:

  *  duckdb
  *  python

```