Add external tests of Bacalhau CLI
This commit adds an external testing framework that runs the Bacalhau
CLI and asserts on the output and exit codes produced. These sorts of
tests are useful checks that what we are running in production
actually behaves the way we expect, and it should be possible to have
more confidence in these test results then in running Go-based tests,
which can inherit a lot of configuration and setup.