Improve pagination and filtering of job list command. (#3344)
Currently we allow users to define limit query params when listing jobs,
but we always return an empty next_token (offset), which prevents users
from actual pagination and querying the next set of jobs.

This PR moves label filtering into the jobstore, and introduces a
default sort order so that the offset/limit make sense. This allows for
proper pagination from the CLI. The next-token is now used by the
jobstore, but it only encodes the offset (e.g. it is the offset). The
next-token should really encode more information so that it can be
re-constituted with the correct parameters. Without this (as currently)
it relies on the same sort being applied for each request.