Make publisher optional in CLI (#3417)
Currently we default to IPFS for the publisher if one is not specified.
This commit changes this to allow no publisher to be specified. It is
expected that this will allow us to run jobs with no publisher, and also
to have the requester node choose which publisher to use.