Require NATS clients to be authenticated with a server secret token
This commit adds simple authentication to our NATS based networking.
Now, NATS servers will self-generate an auth token and print this as
part of their output config, and clients will need to connect with
this auth token. Clients that connect with no or wrong tokens will be
denied access (and will fail on boot).

Note that this doesn't add any *authz*, so connected clients can still
read and write the full contents of the pubsub as desired.