Adds support for specifying a default publisher on the server
As it is no longer mandatory to have a publisher defined for a job
(either by the user, or using the default) there is a chance of
functionality that may leave a gap.

To allow for the old behaviour to continue as a short-term measure, this
commit introduces the `--default-publisher` option for the devstack and
requester/compute node (launched by serve).

Using the `--default-publisher` flag is similar to specifying a
publisher, and so to continue the same behaviour as we currently have
where the CLI has a default, we would start the node with
`--default-publisher ipfs`.

This has been implemented for bother requester and orchestrator (v1 and
v2) job submissions.