Fix too-speedy shutdown in devstack
Currently the main() func and the devstack are sharing a cancellable
context, and when it is cancelled by SIGINT, the main function exits
before the deferred tasks in devstack can complete. This resulted in
devstack not running a second time because of the existing pid or port
(or both) files.

Now main has its own context which is cancelled only when the job is
complete (or cancelled from SIGINT). This gives devstack enough time to
shutdown.