Pre-allocate free ports to avoid re-allocation
Currently it's possible to ask for a free port, and then because it
isn't used immediately, for it to be re-allocated for use by another
service. This leads to errors where ports are already in use when code
that thinks it owns a port comes to use it.

This commit pre-allocates the required number of free ports in the
devstack setup to ensure there are no clashes due to slow starting
servers.