Adds a 'local' publisher (#3431)
Adds a new publisher type that allows for publishing to the local disk
of the compute node. This allows users to test out the publisher
functionality without needing to set up a remote storage service making
it easier to get started with Bacalhau. Currently this implemention is
does not include features such as content lifetimes, or secure access,
and so is not recommended for production use yet. This publisher is
useful for testing and development purposes.

The local publisher comprises two parts. The publisher is responsible
for taking the output from a job, compressing it and moving it to a
location where it can be access by the second part. This part is the
HTTP server which delivers the content back to the caller.

By default the HTTP server listens on port 6001, but this can be changed
by setting the `--local-publisher-port` flag on the compute node. The
HTTP server will serve the content from the directory specified by the
`local-publisher-directory` flag. If this is not set, the HTTP server
will serve the content from a subdirectory of the configured bacalhau
storage directory. Finally the `--local-publisher-address` flag can be
used to specify the address that the HTTP server listens on. The default
for this varies by environment, in test and development environments it
uses `localhost`, in production environments it uses `public` to obtain
a public address. Of course, you can set these values in config in
preference to using the command line flags if the defaults are not
suitable.