Replace freeport dependency with modified version
Currently we use phayes/freeport, but it has a couple of issues and
isn't actively maintained.  As it isn't a large amount of code, it's
imported and modified to provide free ports across all interfaces, not
just the 127.0.0.1 address.

This also leaves us room to extend the package and add per-address
free-port lookups.