Remove invalid or expired auth tokens
This commit allows the Bacalhau CLI to automatically remove tokens
that the server says are not valid, e.g. because they have expired or
because the server no longer recognises the signing authority.

In particular, this process is transparent to the rest of the system -
the authentication happens as part of a normal API call, and the API
call doesn't complete until the authentication flow is done. Naturally
this is an opt-in for anyone using the Bacalhau codebase as a library
because this might not be desired.

This involved splitting the transport parts of the API out from the
domain-specific parts so that the authentication code could monitor
for invalid auth tokens and invoke the authentication flow.