Make it easier to build _just_ the binary
Currently if you want to build the bacalhau binary, you also have to
build the webui and plugins as well, even if you've no intention of
using them.

This change adds:

Just builds the bacalhau binary (assuming webui is available already for
embedding).

Builds the webui ready for embedding in the binary

The existing step now calls binary-web and then binary.