Persistent node store (#3503)
Implements a NATS KV backed store for Node Information to persist node
information across requester node restarts. Currently we only have an
in-memory store, and as a result we lose all knowledge of any compute
nodes when we restart the requester.

The new persistent store is the first step in providing more effective
node information and being able to track compute nodes as they become
inaccessible to the cluster for short periods of time. The KV Store
depends on NATS JetStream and so there have been changes to the set up
of the NAT Server, and a first pass at using a test NATS server in
Bacalhau tests (cf
https://github.com/bacalhau-project/bacalhau/blob/1c1662c595dfacbcbdb124bd4bd4cef3c6bdd6de/pkg/routing/kvstore/kvstore_test.go#L33-L40)

TIme metrics and spans are available for both stores (inmemory for
libp2p and kvstore for NATS) although it is expected that the inmemory
store eventually be removed.

These changes did involve modifications to how the network transport
performed pubsub for adding newly discovered node, only because the
transport depends on the store, but this particular store depends on the
transport and so construction is now a two-step process.

## Summary by CodeRabbit

- **New Features**
- Introduced network storage directory configuration for enhanced data
management.
- Enhanced node discovery process for improved reliability and
efficiency.
- Implemented new logging and node information management features in
transport layers.
- Added key-value store functionality for robust node information
storage and retrieval.
- Introduced metrics and tracing for node management operations,
enhancing monitoring and debugging capabilities.

- **Bug Fixes**
- Adjusted node discovery logic to ensure accurate node selection and
information retrieval.

- **Refactor**
- Unified node information storage interfaces and methods across
different components for consistency and maintainability.
- Streamlined configuration and initialization processes for network and
node components.