Adds support for filtering of the List command in NodeInfoStores
Currently when listing node info in the NodeInfoStores, it is all or
nothing, and there is no way to narrow down the list.

This PR adds support for passing varargs of filter functions so that we
can do things like easily filter by approval state, or GPU count, or
specific labels.  It is expected that these filter functions will be
pre-defined rather than expecting them to be constructed as required.